Taking the yelp dataset challenge reviews .json file, this code will create correlations between each business's star reviews and text sentiment.
Taking the yelp dataset challenge reviews .json file, this code will
create correlations between each business's star reviews and text
sentiment. The text sentiment is calculated using the default settings
of textblob. The code will need to be changed to see the whole dataset.
Right now it only loads the first 21 entries. Remove [0-20] in for d in
reviewData[0:20] to make correlations over the whole dataset. This will
take a long time, as the review dataset is massive.